requirements.txt
pydantic~=2.11.5
python-dotenv~=1.1.0
aiogram~=3.20.0.post0